Return $path without URL encoding in Iceberg
